Add image build, test and release workflow
Build a multi-platform image from the alpine-builder Containerfile
targeting the architectures supported on Linux by Please and its
plugins, test it by running the cc-rules and python-rules test suites on
both platforms, and push the image to GHCR if the tests pass and the
commit is tagged with something that looks like a version number (see
`README.md`).

The entire workflow is sequential and is executed in a single job
because GHA jobs are executed on separate runners, and the only means of
passing information between jobs are either inadequate for ~400MB
container images (job outputs are limited to 1MB) or are costly and/or
difficult to reason about (artifacts persist for longer than the life of
a workflow and have a 500MB per-repo limit, limiting us to one workflow
execution at a time; although unlikely, caches could be evicted between
job executions if the 10GB per-repo limit is breached). It only takes a
few minutes to test each image, though, and the image shouldn't change
very often so this isn't really much of a concern.