Use UID/GID 1001 for `runner` user/group (#3)
When running a GHA job in a container, the runner transparently mounts
the workspace as a volume inside the container. This directory is owned
by the runner image's `runner` user, which on GitHub's official Ubuntu
images has the UID 1001 and GID 1001. Use the same UID and GID for the
`runner` user inside the alpine-builder image to eliminate errors when
attempting to write to the workspace (e.g. by checking out code) inside
the alpine-builder image.