Don't specify a platform for the parent image (#4)
`--platform=$BUILDPLATFORM` forces the parent image to have the same
platform as the native platform on which the image is built, which in
this case is linux/amd64 (because the image is built on GitHub's
`ubuntu-24.04` runner image). The "linux/arm64" variant of the image is
therefore just Alpine Linux x86-64, which silently works - at least,
until you try to run an arm64 binary inside the container - as long as
the host has QEMU installed, because it is able to (potentially very
slowly) emulate the x86-64 instruction set for the container.

Don't force a specific platform in the Containerfile, and trust Docker
to choose a platform-appropriate parent image at build time (which it
does).